fix: more than one pipe working, all tests passed
