--- b
@@ -1,1 +1,0 @@
-banana
