+++ test_minishell.sh
@@ -27,7 +27,8 @@ tests=(
 	"pwd"
 	"cat Makefile | wc | wc -l"
 	"echo "banana" > a > b | ls"
-	"< Makefile wc | ls"
+	"< Makefile wc | ls"	
+	"< Makefile cat | wc | cat -e | cat -e | cat -e | cat -e"
 #  "echo -n hello && echo world"
 #  "cd .. && pwd"
 #  "echo test > tmpfile && cat tmpfile"
@@ -63,9 +64,9 @@ run_test() {
         echo -e "${RED}[KO]${RESET} $CMD"
     fi
         echo -e "${YELLOW}--- minishell ---${RESET}"
-        cat -e "$OUT_MINI"
+#        cat -e "$OUT_MINI"
         echo -e "${YELLOW}--- bash ---${RESET}"
-        cat -e "$OUT_BASH"
+#        cat -e "$OUT_BASH"
         echo -e "${YELLOW}--- status ---${RESET}"
         echo -e "status mini: $STATUS_MINI"
         echo -e "status bash: $STATUS_BASH"
